+++ frontend/public/html/navbar.html
@@ -9,7 +9,7 @@
 	   
 	  <!-- Center: Login/Register Links -->
 	  <div class="flex space-x-4">
-		<a href="/dashboard" id="dashboard-link" class="hover:text-white text-md font-bold px-3 py-2 hidden" data-link="true">Play Game</a>
+		<a href="/dashboard" id="dashboard-link" class="hover:text-white text-md font-bold px-3 py-2 hidden" data-link="true">Jungle Pong</a>
 	  </div>
 	  
 	  <!-- Right: Logout -->
