+++ frontend/public/html/userList.html
@@ -1,5 +1,5 @@
 <div class="mx-auto mt-10 max-w-2xl">
-  <h2 class="mb-6 text-2xl font-bold">All My Friends Are Dead</h2>
+  <h2 class="mb-6 text-2xl font-bold">Currently showing ALL users</h2>
 
   <ul id="user-list" class="rounded-lg bg-white shadow-md">
     <li class="p-4 text-center text-gray-500">Loading users...</li>
