+++ frontend/public/html/userMenu.html
@@ -14,6 +14,13 @@
       My Profile
     </a>
     <a
+      href="/friends"
+      class="block rounded-lg px-3 py-2 hover:bg-(--color-primary-light)"
+      data-link="true"
+    >
+      My Friends
+    </a>
+    <a
       href="/users"
       class="block rounded-lg px-3 py-2 hover:bg-(--color-primary-light)"
       data-link="true"
@@ -22,4 +29,3 @@
     </a>
   </nav>
 </div>
-
