+++ frontend/public/html/userProfile.html
@@ -169,7 +169,7 @@
   <!-- Avatar Selection Modal -->
   <div
     id="avatar-modal"
-    class="modal fixed inset-0 z-50 hidden items-center justify-center"
+    class="modal fixed inset-0 z-50 hidden items-center justify-center backdrop-blur-sm"
   >
     <div
       class="mx-4 max-w-md rounded-lg border-2 border-(--color-primary) bg-white p-6 shadow-xl"
